+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### VNEXT
 
 * Incompatible fragments are now properly filtered [PR #470](https://github.com/apollographql/graphql-tools/pull/470)
+* Remove `isTypeOf` guards from merged schemas [PR #484](https://github.com/apollographql/graphql-tools/pull/484)
 
 ### v2.7.1